[extension] Remove Settings.ModuleInfo and move to service.Host hidden method
